added output files, error catching for failed images

+++ astrom_solve.py
@@ -50,6 +50,7 @@ $ sh ./getindex.sh
 """
 
 import argparse
+import fitsio
 import glob
 import os
 import warnings
@@ -59,6 +60,7 @@ import numpy as np
 from astropy.utils.exceptions import AstropyWarning
 from astropy.io import fits
 from astropy.table import Table
+#from astropy.wcs import WCS
 from matplotlib import pyplot as plt
 
 import astrom_config
@@ -84,6 +86,9 @@ def residual_plot(image):
     norm_flux = img_sources['FLUX'] / np.mean(img_sources['FLUX'])
     sym_size = [0 if f < 0 else f for f in norm_flux]
     
+    # setup axes to be WCS aligned - FUTURE WORK
+    #plt.axes(projection=wcs)
+    
     # all detected sources
     plt.scatter(img_sources['X'],img_sources['Y'],s=sym_size,color='slateblue',label='Extracted sources')
     
@@ -105,6 +110,7 @@ def residual_plot(image):
     plt.title(image+ '\n Residuals (N='+str(len(cat_sources))+'): max = '+f'{max_residual:.3f} pix, mean = ' + f'{mean_residual:.3f} pix')
     plt.legend(loc='lower right',ncols=2)
     plt.tight_layout()
+    plt.gca().set_aspect('equal')
     plt.savefig(image.replace('.fits','_solved.png'),dpi=150)
     plt.clf()
     
@@ -114,7 +120,10 @@ def residual_plot(image):
 
 def astrom_solve(image,params):
 
-    # retrieve header keywords
+    # retrieve header keywords for processing and summary file
+    dat_obs = fits.getval(image, 'DATE-OBS', ext=0)     # UT date of observation
+    filt = fits.getval(image, 'FILTER', ext=0)          # Filter
+    obj = fits.getval(image, 'OBJECT', ext=0)           # Object
     ra = fits.getval(image, 'RA', ext=0)    # right ascension
     dec = fits.getval(image, 'DEC', ext=0)  # declination
     pix_scale = params['pix_scale']         # unbinned pixel scale
@@ -149,21 +158,59 @@ def astrom_solve(image,params):
     #   -N              new fits file output
     #   -W              wcs file
     #
-    solve_field = 'solve-field "' + image + '" --scale-low ' + str(scale_low) + ' --scale-high ' + str(scale_high) + ' --scale-units arcsecperpix -O --ra ' + ra + ' --dec ' + dec + ' --radius ' + str(params['search_radius']) + ' --no-plots -t ' + str(params['poly_n']) + ' --no-verify -k %s.xy -S none -M none -R none --temp-axy -U none -N %s_solved.fits'
+    solve_field = 'solve-field "' + image + '" --scale-low ' + str(scale_low) + ' --scale-high ' + str(scale_high) + ' --scale-units arcsecperpix -O --ra ' + ra + ' --dec ' + dec + ' --radius ' + str(params['search_radius']) + ' --no-plots -t ' + str(params['poly_n']) + ' --no-verify -k %s.xy -S %s_solved -M none -R none --temp-axy -U none -N %s_solved.fits'
 
     # run solve-field
     os.system(solve_field)
-        
-    # create plot of residuals
-    if do_plot:
-        residual_plot(image)
+  
+    # retrive WCS solution and image rotation - FUTURE WORK
+    #head = fitsio.read_header(image.replace('.fits','.wcs'))
+    #wcs = WCS(head)
+    #cd = wcs.wcs.cd
+    #theta = np.arctan2(-cd[0, 1], cd[0, 0])
+    #rotation_angle = np.degrees(theta)
+    
+    # generate output if solve was successful
+    if os.path.isfile(image.replace('.fits','_solved')):
 
-    # clean up output files from solve-field
-    if cleanup:
-        os.remove(image.replace('.fits','.corr'))
-        os.remove(image.replace('.fits','.wcs'))
-        os.remove(image.replace('.fits','.xy'))
+        # retrieve results
+        n_sources = len(Table.read(image.replace('.fits','.xy')))
+        with warnings.catch_warnings():
+            warnings.simplefilter('ignore', category=AstropyWarning)
+            catalog = Table.read(image.replace('.fits','.corr'))
+
+        # residuals on astrometric solution
+        dx = (catalog['field_x'] - catalog['index_x'])*1
+        dy = (catalog['field_y'] - catalog['index_y'])*1
+        max_res = np.sqrt(sorted(dx)[-1]**2 + sorted(dy)[-1]**2)
+        mean_res = np.mean(np.sqrt(np.array(dx)**2 + np.array(dy)**2))
+
+        # summary data
+        solve_dat = (image, dat_obs, ra, dec, filt, obj, n_sources, len(catalog), max_res, mean_res)
+             
+        # create plot of residuals
+        if do_plot:
+            residual_plot(image)
+
+        # clean up output files from solve-field
+        if cleanup:
+            os.remove(image.replace('.fits','.corr'))
+            os.remove(image.replace('.fits','.wcs'))
+            os.remove(image.replace('.fits','.xy'))
+            os.remove(image.replace('.fits','_solved'))
+
+        return solve_dat
+    
+    # bookeeping if solve failed
+    else:
+        # construct summary data for failed images
+        failed_dat = (image, dat_obs, ra, dec, filt, obj, 0, 0, 0.0, 0.0)
         
+        with open(image.replace('.fits','_FAILED'), 'w') as f:
+            f.write('Image '+image+' failed to solve')
+        
+        return failed_dat
+    
 ##############################
 # Main Block
 ##############################
@@ -177,9 +224,18 @@ if __name__ == '__main__':
     args = parser.parse_args()
     
     img_path = args.imgpath
+    # add trailing slash / if missing from image path name
+    if img_path[-1] != '/':
+        img_path += '/'
+    
     cleanup = args.cleanup
     do_plot = args.plot
 
+    # astrometry summary table
+    summary = Table(names=('image', 'UT Date','RA', 'Dec', 'filter', 'object', 'N_sources', 'N_catalog', 'Max_pix_residual', 'Mean_pix_residual'), dtype=(str,str,str,str,str,str,int,int,float,float))
+    
+    # list of images that failed to register
+    failed = []
 
     # retrieve list of images, exclude previously solved fits files
     # assumes original file name does not include the string 'solved'
@@ -200,7 +256,34 @@ if __name__ == '__main__':
     
         # solve images
         for im in images:
-            astrom_solve(im,params)
+            summary_row = astrom_solve(im,params)
             
+            # add image data to summary table
+            summary.add_row(summary_row)
+            
+            # record if image failed to solve
+            if os.path.isfile(im.replace('.fits','_FAILED')):
+                failed.append(summary_row)
+            
+        # write all summary data to file
+        summary.write('astrom_summary.txt', format='ascii.fixed_width_two_line', formats={'Max_pix_residual':'0.3f', 'Mean_pix_residual':'0.3f'}, overwrite=True)
+        
+        # overall mean residual for successfully solved images
+        mean_resid = np.mean(summary['Mean_pix_residual'][summary['Mean_pix_residual'] > 0])
+        
+        # print some information to terminal
+        print('-'*80)
+        print('SUMMARY')
+        print('-'*80)
+        print(str(len(summary)) + ' files in path ' + img_path)
+        print(str(len(summary) - len(failed)) + ' files successfully solved')
+        print('Mean residual (pixels) = ' + f'{mean_resid:.3f}')
+        if len(failed) > 0:
+            print(str(len(failed)) + ' files failed')
+            print('   Images that didnt solve:')
+            for f in failed:
+                print('   ',f)
+        print('\nDONE')
+
     else:
         print('No .fits images found in path '+img_path)
